+++ Readme.md
@@ -1,0 +1,62 @@
+# DataHub Charts
+
+Charting library for Datahub
+
+## Installation
+
+If you're using yarn
+```
+yarn add @devinit/charts
+```
+
+If you're still using npm :(
+```
+npm install @devinit/charts --save
+```
+
+## Testing
+```
+npm test
+```
+
+## Usage
+
+### Bar
+
+TODO
+
+### Area
+
+TODO
+
+### Bubble
+
+TODO
+
+### Column
+
+TODO
+
+### Donut
+
+TODO
+
+### Grouped Column
+
+TODO
+
+### Line
+
+TODO
+
+### Pie
+
+TODO
+
+### Stacked Column
+
+TODO
+
+### Tree
+
+TODO
