--- index.js
@@ -1,13 +1,0 @@
-import vb_column from './column';
-import vb_bar from './bar';
-import vb_stacked_column from './stacked_column';
-import vb_donut from './donut';
-import vb_pie from './pie';
-
-module.exports = {
-    vb_column,
-    vb_bar,
-    vb_stacked_column,
-    vb_donut,
-    vb_pie,
-};
